[stream]兼容snakeyaml解析出错 #8823
