+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/trigger/parsers/yamlCheck/YamlSchemaCheck.kt
@@ -89,7 +89,7 @@ class YamlSchemaCheck @Autowired constructor() {
 
     private fun checkYamlSchema(originYaml: String, templateType: TemplateType? = null, isCiFile: Boolean) {
         val loadYaml = try {
-            YamlUtil.toYaml(Yaml().load(originYaml))
+            YamlUtil.loadYamlRetryOnAccident(originYaml)
         } catch (ignored: Throwable) {
             logger.warn("YAML_SCHEMA_CHECK|originYaml=$originYaml", ignored)
             throw YamlFormatException("There may be a problem with your yaml syntax ${ignored.message}")
